refactor(frontend): tighten history public surface
