+++ app/frontend/src/features/history/index.ts
@@ -1,3 +1,6 @@
-export * from './components';
-export * from './composables';
-export * from './services';
+export {
+  GenerationHistory,
+  HistoryRecentResultCard,
+  listResults,
+} from './public';
+export type { ListResultsOptions, ListResultsOutput } from './public';
